Move modal to folder
