+++ src/components/modal/index.tsx
@@ -1,0 +1,1 @@
+export { Modal } from './Modal'
